requirements/sector_rotation.txt
pandas>=2.1
numpy>=1.26
yfinance>=0.2
scipy>=1.11
statsmodels>=0.14
matplotlib>=3.8
PyYAML>=6.0
hmmlearn>=0.3
scikit-learn>=1.3
pytest>=7.4
